DRY up exporter code: Move common code to Dhis2ExporterJob
