+++ lib/tasks/dhis2.rake
@@ -2,7 +2,7 @@ namespace :dhis2 do
   desc "Export indicator data of each facility to Bangladesh DHIS2"
   task bangladesh_export: :environment do
     FacilityBusinessIdentifier.dhis2_org_unit_id.each do |facility_identifier|
-      Dhis2::BangladeshExporterJob.perform_async(facility_identifier.id, 24)
+      Dhis2::Dhis2ExporterJob.perform_async(facility_identifier.id, 24)
     end
   end
 
@@ -16,7 +16,7 @@ namespace :dhis2 do
   desc "Export data of each facility to Ethiopia DHIS2"
   task ethiopia_export: :environment do
     FacilityBusinessIdentifier.dhis2_org_unit_id.each do |facility_identifier|
-      Dhis2::EthiopiaExporterJob.perform_async(facility_identifier.id, 24)
+      Dhis2::Dhis2ExporterJob.perform_async(facility_identifier.id, 24)
     end
   end
 end
