+++ spec/jobs/dhis2/dhis2_exporter_job_spec.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+describe Dhis2::Dhis2ExporterJob do
+  before do
+    # Do nothing
+  end
+
+  after do
+    # Do nothing
+  end
+
+  context 'when condition' do
+    it 'succeeds' do
+      pending 'Not implemented'
+    end
+  end
+end
